delete fix

+++ src/main/java/edu/harvard/dbmi/avillach/dictionaryetl/dataset/DatasetController.java
@@ -125,7 +125,9 @@ public class DatasetController {
             datasetMetadataRepository.findByDatasetId(datasetId).forEach(dm -> {
                 datasetMetadataRepository.delete(dm);
             });
-            facetRepository.delete(facetRepository.findByName(datasetRef).get());
+            if (facetRepository.findByName(datasetRef).isPresent()) {
+                facetRepository.delete(facetRepository.findByName(datasetRef).get());
+            }
             consentRepository.findByDatasetId(datasetId).forEach(consent -> {
                 consentRepository.delete(consent);
             });
